Attempt to add salting-and-hashing code
The sha3() call within the saltAndHash method I added definitely won't compile (as I have no clue how to access the actual SHA-3 algorithm from within PHP), but that aside, everything else should work, assuming that I, in my great newbyness, didn't butcher anything horribly. And that the database is OK with a new column being written to.

You'll also need to add the same combine-password-with-salt-then-hash code (or, better yet, a call to the saltAndHash method in this file) in whatever code you have to change passwords on existing accounts.